functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.0.0
requests>=2.20.0
PyJWT[crypto]>=2.0.0
python-dotenv>=1.0.0